Refactored code for stack resoruces
